major refactor - corpus.py
TextGrid handling was moved to datapipes/. This code now only deals with
tar.gz files.

speechcorpus.py was renamed to corpus.py. It loads audio files only as
needed and does not retain them in memory. It only contains utility
functions for loading parts of the corpus.

soundsheaf.py was removed. The same concepts will probably be reused,
but as written, it doesn't work well for model-building.