new classes for loading clipper data
This separates between Clipper's format and data. The clipper format is
an association between audio files, transcript files, ponysorter files,
and audacity files. All of these things are linked by a label reference,
which is a normalized version of the audacity label. The actual file
locations are separate from this structure, as are the name choices and
the method for linking names across files.